+++ src/ios/syncHost.ts
@@ -218,7 +218,9 @@ class ViewController: UIViewController {
         additionalSafeAreaInsets = .zero
         view.insetsLayoutMarginsFromSafeArea = false
         view.preservesSuperviewLayoutMargins = false
-        viewRespectsSystemMinimumLayoutMargins = false
+        if #available(iOS 15.0, *) {
+            viewRespectsSystemMinimumLayoutMargins = false
+        }
     }
 
     override func viewDidLayoutSubviews() {
@@ -280,6 +282,7 @@ function getDevViewControllerSwift(): string {
 import Lynx
 import tamerdevclient
 import tamerinsets
+import tamersystemui
 
 class ViewController: UIViewController {
     private var lynxView: LynxView?
@@ -292,7 +295,9 @@ class ViewController: UIViewController {
         additionalSafeAreaInsets = .zero
         view.insetsLayoutMarginsFromSafeArea = false
         view.preservesSuperviewLayoutMargins = false
-        viewRespectsSystemMinimumLayoutMargins = false
+        if #available(iOS 15.0, *) {
+            viewRespectsSystemMinimumLayoutMargins = false
+        }
         setupLynxView()
         setupDevClientModule()
     }
@@ -309,7 +314,7 @@ class ViewController: UIViewController {
         TamerInsetsModule.reRequestInsets()
     }
 
-    override var preferredStatusBarStyle: UIStatusBarStyle { TamerPreferredStatusBar.style }
+    override var preferredStatusBarStyle: UIStatusBarStyle { SystemUIModule.statusBarStyleForHost }
 
     private func setupLynxView() {
         let size = fullscreenBounds().size